feat: add paste rules
